Fixed error prepopulated_fields without max_length
Prevented an attribute error when specifying a field without a max_length attribute.
In the absence of the attribute, an error was raised while the code was reading like it
should evaluate the left hand of the `or` and use the value 50. Using `getattr` will 
prevent the error which was occurring on numeric field for exemple.